Edit : total_price field in order model

+++ pizza/orders/models.py
@@ -62,7 +62,7 @@ class CartItem(models.Model):
 
 class Order(models.Model):
     items = models.ManyToManyField(CartItem, related_name="orders")
-    total_price = models.IntegerField(default=0)
+    total_price = models.FloatField(default=0)
     costumer = models.ForeignKey(User, on_delete=models.CASCADE, related_name="orders", blank=True, default=0)
 
     def __str__(self):
